Wire team stage/drop pipeline to single-prefab factories and SO data.
MonsterSpawner uses MonsterFactory.Create; ItemDropManager spawns via ItemFactory; Stage tracks kills only; Main/Test scenes get Facades and Excel import menus.

+++ Assets/01.Script/06.Monster/MonsterFacade.cs
@@ -108,7 +108,7 @@ public class MonsterFacade : MonoBehaviour
     {
         if (itemDropFacade != null)
         {
-            int dropTableId = info.DropTableId > 0 ? info.DropTableId : info.MonsterId;
+            int dropTableId = info.DropTableId > 0 ? info.DropTableId : (int)info.MonsterId;
             itemDropFacade.RequestDrop(dropTableId, info.Position);
         }
